refactor: move `extract_messages_from_code` to a separate utils file
In order to avoid circular imports.
